Add support for 32-bit entities

+++ include/ecs/ComponentManagerImpl.hh
@@ -145,7 +145,7 @@ namespace ecs
 		Assert(entCompMasks.size() > e.Index(), "entity does not have a component mask");
 
 		auto &compMask = entCompMasks.at(e.Index());
-		for (uint64 i = 0; i < componentPools.size(); ++i)
+		for (size_t i = 0; i < componentPools.size(); ++i)
 		{
 			if (compMask[i])
 			{
